Add ElicitSchemaValidator utility
