+++ src/main/java/com/amannmalik/mcp/spi/ElicitRequest.java
@@ -1,5 +1,6 @@
 package com.amannmalik.mcp.spi;
 
+import com.amannmalik.mcp.util.ElicitSchemaValidator;
 import com.amannmalik.mcp.util.ValidationUtil;
 import jakarta.json.JsonObject;
 
@@ -9,7 +10,7 @@ public record ElicitRequest(String message, JsonObject requestedSchema, JsonObje
             throw new IllegalArgumentException("message and requestedSchema are required");
         }
         this.message = ValidationUtil.requireClean(message);
-        ValidationUtil.requireElicitSchema(requestedSchema);
+        ElicitSchemaValidator.requireElicitSchema(requestedSchema);
         this.requestedSchema = requestedSchema;
         ValidationUtil.requireMeta(_meta);
         this._meta = _meta;
